Issue-339: Added MainFragment.
